api: split EventService helpers into modules
