+++ crates/api/src/event_service.rs
@@ -1,10 +1,7 @@
 //! Unified live event stream service.
 
-use std::collections::BTreeSet;
-use std::net::IpAddr;
 use std::pin::Pin;
 use std::sync::{Arc, Mutex};
-use std::time::{SystemTime, UNIX_EPOCH};
 
 use tokio::sync::{broadcast, oneshot};
 use tokio_stream::Stream;
@@ -13,18 +10,30 @@ use tokio_stream::wrappers::{BroadcastStream, errors::BroadcastStreamRecvError};
 use tonic::{Request, Response, Status};
 use tracing::debug;
 
-use crate::peer_types::{
-    PeerManagerCommand, PolicyEvent, SessionEvent, SessionLifecycleEvent,
-    SessionLifecycleEventType, SessionNotificationEvent, SessionNotificationEventType,
+mod convert;
+mod dataplane;
+mod filters;
+
+use convert::{
+    policy_event_to_bgp_event, route_event_to_bgp_event, session_event_to_bgp_event,
+    stream_lag_bgp_event,
 };
+use dataplane::spawn_dataplane_poller;
+#[cfg(test)]
+use dataplane::{DataplaneSummary, dataplane_summaries};
+use filters::{
+    parse_list_policy_events_filter, parse_list_session_events_filter, parse_watch_events_filter,
+};
+
+use crate::peer_types::{PeerManagerCommand, SessionEvent};
 use crate::proto;
-use crate::rib_service::{
-    BlackholeDiscardSnapshotFn, FibRouteSnapshotFn, parse_route_event_prefix_filter,
-    route_event_afi_filter, route_event_to_proto,
-};
-use rustbgpd_rib::{RibUpdate, RouteEventType};
+use crate::rib_service::{BlackholeDiscardSnapshotFn, FibRouteSnapshotFn};
+use rustbgpd_rib::RibUpdate;
+#[cfg(test)]
+use rustbgpd_rib::RouteEventType;
 use rustbgpd_telemetry::BgpMetrics;
-use rustbgpd_wire::{Afi, Prefix};
+#[cfg(test)]
+use rustbgpd_wire::Prefix;
 
 #[cfg(not(test))]
 const DATAPLANE_EVENT_POLL_INTERVAL: std::time::Duration = std::time::Duration::from_secs(1);
@@ -137,766 +146,6 @@ impl EventService {
             self.fib_route_snapshot.clone(),
         );
         rx
-    }
-}
-
-#[derive(Clone)]
-struct WatchEventsFilter {
-    categories: BTreeSet<i32>,
-    event_types: BTreeSet<i32>,
-    peer: Option<IpAddr>,
-    afi: Option<Afi>,
-    prefix: Option<Prefix>,
-}
-
-impl WatchEventsFilter {
-    fn matches_route_event(&self, event: &rustbgpd_rib::RouteEvent) -> bool {
-        if !self.wants_route_events() {
-            return false;
-        }
-
-        let event_type = route_event_type_to_bgp_event_type(event.event_type);
-        if !self.event_types.is_empty() && !self.event_types.contains(&(event_type as i32)) {
-            return false;
-        }
-
-        if let Some(afi) = self.afi {
-            match (afi, event.prefix) {
-                (Afi::Ipv4, Prefix::V4(_)) | (Afi::Ipv6, Prefix::V6(_)) => {}
-                _ => return false,
-            }
-        }
-
-        if let Some(peer) = self.peer {
-            let matches_current = event.peer == Some(peer);
-            let matches_previous = event.previous_peer == Some(peer);
-            if !matches_current && !matches_previous {
-                return false;
-            }
-        }
-
-        if let Some(prefix) = self.prefix
-            && event.prefix != prefix
-        {
-            return false;
-        }
-
-        true
-    }
-
-    fn wants_route_events(&self) -> bool {
-        (self.categories.is_empty()
-            || self
-                .categories
-                .contains(&(proto::EventCategory::Route as i32)))
-            && self.event_types_match_route_category()
-    }
-
-    fn wants_session_events(&self) -> bool {
-        (self.categories.is_empty()
-            || self
-                .categories
-                .contains(&(proto::EventCategory::Session as i32)))
-            && self.afi.is_none()
-            && self.prefix.is_none()
-            && self.event_types_match_session_category()
-    }
-
-    fn wants_policy_events(&self) -> bool {
-        let policy_category_requested = self
-            .categories
-            .contains(&(proto::EventCategory::Policy as i32));
-        let policy_type_allowed = self.event_types_match_policy_category();
-        let policy_selected = policy_category_requested
-            || (self.categories.is_empty() && !self.event_types.is_empty() && policy_type_allowed);
-        policy_selected && self.afi.is_none() && self.prefix.is_none() && policy_type_allowed
-    }
-
-    fn wants_dataplane_events(&self) -> bool {
-        let dataplane_category_requested = self
-            .categories
-            .contains(&(proto::EventCategory::Dataplane as i32));
-        let dataplane_type_allowed = self.event_types_match_dataplane_category();
-        let dataplane_selected = dataplane_category_requested
-            || (self.categories.is_empty()
-                && !self.event_types.is_empty()
-                && dataplane_type_allowed);
-        dataplane_selected
-            && self.peer.is_none()
-            && self.afi.is_none()
-            && self.prefix.is_none()
-            && dataplane_type_allowed
-    }
-
-    fn matches_session_event(&self, event: &SessionEvent) -> bool {
-        if !self.wants_session_events() {
-            return false;
-        }
-
-        let (event_type, event_peer) = match event {
-            SessionEvent::Lifecycle(event) => (
-                session_lifecycle_event_type_to_bgp_event_type(event.event_type),
-                event.peer,
-            ),
-            SessionEvent::Notification(event) => (
-                session_notification_event_type_to_bgp_event_type(event.event_type),
-                event.peer,
-            ),
-        };
-        if !self.event_types.is_empty() && !self.event_types.contains(&(event_type as i32)) {
-            return false;
-        }
-
-        if let Some(peer) = self.peer
-            && event_peer != peer
-        {
-            return false;
-        }
-
-        true
-    }
-
-    fn matches_policy_event(&self, event: &PolicyEvent) -> bool {
-        if !self.wants_policy_events() {
-            return false;
-        }
-
-        if !self.event_types.is_empty()
-            && !self
-                .event_types
-                .contains(&(proto::BgpEventType::PolicyChanged as i32))
-        {
-            return false;
-        }
-
-        if let Some(peer) = self.peer
-            && event.peer != Some(peer)
-        {
-            return false;
-        }
-
-        true
-    }
-
-    fn event_types_match_route_category(&self) -> bool {
-        self.event_types.is_empty()
-            || self.event_types.iter().any(|event_type| {
-                matches!(
-                    proto::BgpEventType::try_from(*event_type),
-                    Ok(proto::BgpEventType::RouteAdded
-                        | proto::BgpEventType::RouteWithdrawn
-                        | proto::BgpEventType::RouteBestChanged
-                        | proto::BgpEventType::StreamLagged)
-                )
-            })
-    }
-
-    fn event_types_match_session_category(&self) -> bool {
-        self.event_types.is_empty()
-            || self.event_types.iter().any(|event_type| {
-                matches!(
-                    proto::BgpEventType::try_from(*event_type),
-                    Ok(proto::BgpEventType::SessionStateChanged
-                        | proto::BgpEventType::SessionEstablished
-                        | proto::BgpEventType::SessionLost
-                        | proto::BgpEventType::PeerEnabled
-                        | proto::BgpEventType::PeerDisabled
-                        | proto::BgpEventType::NotificationSent
-                        | proto::BgpEventType::NotificationReceived
-                        | proto::BgpEventType::StreamLagged)
-                )
-            })
-    }
-
-    fn event_types_match_policy_category(&self) -> bool {
-        self.event_types.is_empty()
-            || self.event_types.iter().any(|event_type| {
-                matches!(
-                    proto::BgpEventType::try_from(*event_type),
-                    Ok(proto::BgpEventType::PolicyChanged)
-                )
-            })
-    }
-
-    fn event_types_match_dataplane_category(&self) -> bool {
-        self.event_types.is_empty()
-            || self.event_types.iter().any(|event_type| {
-                matches!(
-                    proto::BgpEventType::try_from(*event_type),
-                    Ok(proto::BgpEventType::DataplaneStatusChanged)
-                )
-            })
-    }
-}
-
-struct SessionEventHistoryFilter {
-    peer: Option<IpAddr>,
-    event_types: BTreeSet<SessionLifecycleEventType>,
-    limit: usize,
-}
-
-struct PolicyEventHistoryFilter {
-    peer: Option<IpAddr>,
-    limit: usize,
-}
-
-fn route_event_type_to_bgp_event_type(event_type: RouteEventType) -> proto::BgpEventType {
-    match event_type {
-        RouteEventType::Added => proto::BgpEventType::RouteAdded,
-        RouteEventType::Withdrawn => proto::BgpEventType::RouteWithdrawn,
-        RouteEventType::BestChanged => proto::BgpEventType::RouteBestChanged,
-    }
-}
-
-fn bgp_event_type_to_session_event_type(
-    event_type: proto::BgpEventType,
-) -> Option<SessionLifecycleEventType> {
-    match event_type {
-        proto::BgpEventType::SessionStateChanged => Some(SessionLifecycleEventType::StateChanged),
-        proto::BgpEventType::SessionEstablished => Some(SessionLifecycleEventType::Established),
-        proto::BgpEventType::SessionLost => Some(SessionLifecycleEventType::Lost),
-        proto::BgpEventType::PeerEnabled => Some(SessionLifecycleEventType::PeerEnabled),
-        proto::BgpEventType::PeerDisabled => Some(SessionLifecycleEventType::PeerDisabled),
-        proto::BgpEventType::Unspecified
-        | proto::BgpEventType::RouteAdded
-        | proto::BgpEventType::RouteWithdrawn
-        | proto::BgpEventType::RouteBestChanged
-        | proto::BgpEventType::NotificationSent
-        | proto::BgpEventType::NotificationReceived
-        | proto::BgpEventType::PolicyChanged
-        | proto::BgpEventType::DataplaneStatusChanged
-        | proto::BgpEventType::StreamLagged => None,
-    }
-}
-
-fn parse_category_filter(categories: &[i32]) -> Result<BTreeSet<i32>, Status> {
-    let mut parsed = BTreeSet::new();
-    for category in categories {
-        let category = proto::EventCategory::try_from(*category)
-            .map_err(|_| Status::invalid_argument("unknown event category"))?;
-        match category {
-            proto::EventCategory::Route
-            | proto::EventCategory::Session
-            | proto::EventCategory::Policy
-            | proto::EventCategory::Dataplane => {
-                parsed.insert(category as i32);
-            }
-            proto::EventCategory::Unspecified => {
-                return Err(Status::invalid_argument(
-                    "EVENT_CATEGORY_UNSPECIFIED is not a valid filter",
-                ));
-            }
-        }
-    }
-    Ok(parsed)
-}
-
-fn parse_event_type_filter(event_types: &[i32]) -> Result<BTreeSet<i32>, Status> {
-    let mut parsed = BTreeSet::new();
-    for event_type in event_types {
-        let event_type = proto::BgpEventType::try_from(*event_type)
-            .map_err(|_| Status::invalid_argument("unknown event type"))?;
-        match event_type {
-            proto::BgpEventType::RouteAdded
-            | proto::BgpEventType::RouteWithdrawn
-            | proto::BgpEventType::RouteBestChanged
-            | proto::BgpEventType::SessionStateChanged
-            | proto::BgpEventType::SessionEstablished
-            | proto::BgpEventType::SessionLost
-            | proto::BgpEventType::PeerEnabled
-            | proto::BgpEventType::PeerDisabled
-            | proto::BgpEventType::NotificationSent
-            | proto::BgpEventType::NotificationReceived
-            | proto::BgpEventType::PolicyChanged
-            | proto::BgpEventType::DataplaneStatusChanged
-            | proto::BgpEventType::StreamLagged => {
-                parsed.insert(event_type as i32);
-            }
-            proto::BgpEventType::Unspecified => {
-                return Err(Status::invalid_argument(
-                    "BGP_EVENT_TYPE_UNSPECIFIED is not a valid filter",
-                ));
-            }
-        }
-    }
-    Ok(parsed)
-}
-
-fn parse_session_event_type_filter(
-    event_types: &[i32],
-) -> Result<BTreeSet<SessionLifecycleEventType>, Status> {
-    let mut parsed = BTreeSet::new();
-    for event_type in event_types {
-        let event_type = proto::BgpEventType::try_from(*event_type)
-            .map_err(|_| Status::invalid_argument("unknown event type"))?;
-        if event_type == proto::BgpEventType::Unspecified {
-            return Err(Status::invalid_argument(
-                "BGP_EVENT_TYPE_UNSPECIFIED is not a valid filter",
-            ));
-        }
-        let Some(session_type) = bgp_event_type_to_session_event_type(event_type) else {
-            return Err(Status::invalid_argument(
-                "ListSessionEvents only supports session event types",
-            ));
-        };
-        parsed.insert(session_type);
-    }
-    Ok(parsed)
-}
-
-fn parse_policy_event_type_filter(event_types: &[i32]) -> Result<(), Status> {
-    for event_type in event_types {
-        let event_type = proto::BgpEventType::try_from(*event_type)
-            .map_err(|_| Status::invalid_argument("unknown event type"))?;
-        match event_type {
-            proto::BgpEventType::PolicyChanged => {}
-            proto::BgpEventType::Unspecified => {
-                return Err(Status::invalid_argument(
-                    "BGP_EVENT_TYPE_UNSPECIFIED is not a valid filter",
-                ));
-            }
-            proto::BgpEventType::RouteAdded
-            | proto::BgpEventType::RouteWithdrawn
-            | proto::BgpEventType::RouteBestChanged
-            | proto::BgpEventType::SessionStateChanged
-            | proto::BgpEventType::SessionEstablished
-            | proto::BgpEventType::SessionLost
-            | proto::BgpEventType::PeerEnabled
-            | proto::BgpEventType::PeerDisabled
-            | proto::BgpEventType::NotificationSent
-            | proto::BgpEventType::NotificationReceived
-            | proto::BgpEventType::DataplaneStatusChanged
-            | proto::BgpEventType::StreamLagged => {
-                return Err(Status::invalid_argument(
-                    "ListPolicyEvents only supports policy event types",
-                ));
-            }
-        }
-    }
-    Ok(())
-}
-
-fn session_lifecycle_event_type_to_bgp_event_type(
-    event_type: SessionLifecycleEventType,
-) -> proto::BgpEventType {
-    match event_type {
-        SessionLifecycleEventType::StateChanged => proto::BgpEventType::SessionStateChanged,
-        SessionLifecycleEventType::Established => proto::BgpEventType::SessionEstablished,
-        SessionLifecycleEventType::Lost => proto::BgpEventType::SessionLost,
-        SessionLifecycleEventType::PeerEnabled => proto::BgpEventType::PeerEnabled,
-        SessionLifecycleEventType::PeerDisabled => proto::BgpEventType::PeerDisabled,
-    }
-}
-
-fn session_notification_event_type_to_bgp_event_type(
-    event_type: SessionNotificationEventType,
-) -> proto::BgpEventType {
-    match event_type {
-        SessionNotificationEventType::Sent => proto::BgpEventType::NotificationSent,
-        SessionNotificationEventType::Received => proto::BgpEventType::NotificationReceived,
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-struct DataplaneSummary {
-    source: &'static str,
-    installed: u64,
-    rejected: u64,
-    failed: u64,
-}
-
-impl DataplaneSummary {
-    fn changed_from(self, previous: Self) -> bool {
-        self.installed != previous.installed
-            || self.rejected != previous.rejected
-            || self.failed != previous.failed
-    }
-}
-
-fn count_blackhole_statuses(rows: &[proto::BlackholeDiscard]) -> DataplaneSummary {
-    let mut summary = DataplaneSummary {
-        source: "blackhole",
-        installed: 0,
-        rejected: 0,
-        failed: 0,
-    };
-    for row in rows {
-        match proto::BlackholeDiscardState::try_from(row.state) {
-            Ok(proto::BlackholeDiscardState::Installed) => summary.installed += 1,
-            Ok(proto::BlackholeDiscardState::Rejected) => summary.rejected += 1,
-            Ok(proto::BlackholeDiscardState::Failed) => summary.failed += 1,
-            Ok(proto::BlackholeDiscardState::Unspecified) | Err(_) => {}
-        }
-    }
-    summary
-}
-
-fn count_fib_statuses(rows: &[proto::FibRouteStatus]) -> DataplaneSummary {
-    let mut summary = DataplaneSummary {
-        source: "fib",
-        installed: 0,
-        rejected: 0,
-        failed: 0,
-    };
-    for row in rows {
-        match proto::FibRouteState::try_from(row.state) {
-            Ok(proto::FibRouteState::Installed) => summary.installed += 1,
-            Ok(proto::FibRouteState::Rejected) => summary.rejected += 1,
-            Ok(proto::FibRouteState::Failed) => summary.failed += 1,
-            Ok(proto::FibRouteState::Unspecified) | Err(_) => {}
-        }
-    }
-    summary
-}
-
-fn dataplane_summaries(
-    blackholes: &[proto::BlackholeDiscard],
-    fib_routes: &[proto::FibRouteStatus],
-) -> Vec<DataplaneSummary> {
-    vec![
-        count_blackhole_statuses(blackholes),
-        count_fib_statuses(fib_routes),
-    ]
-}
-
-fn changed_dataplane_summaries(
-    previous: &[DataplaneSummary],
-    current: &[DataplaneSummary],
-) -> Vec<DataplaneSummary> {
-    current
-        .iter()
-        .copied()
-        .filter(|summary| {
-            previous
-                .iter()
-                .find(|candidate| candidate.source == summary.source)
-                .is_none_or(|previous| summary.changed_from(*previous))
-        })
-        .collect()
-}
-
-fn dataplane_summary_to_bgp_event(summary: DataplaneSummary) -> proto::BgpEvent {
-    let severity = if summary.failed > 0 {
-        proto::EventSeverity::Warning
-    } else {
-        proto::EventSeverity::Info
-    };
-    let payload = proto::DataplaneEvent {
-        source: summary.source.to_string(),
-        installed: summary.installed,
-        rejected: summary.rejected,
-        failed: summary.failed,
-    };
-    proto::BgpEvent {
-        timestamp: SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .unwrap_or_default()
-            .as_secs()
-            .to_string(),
-        category: proto::EventCategory::Dataplane as i32,
-        event_type: proto::BgpEventType::DataplaneStatusChanged as i32,
-        severity: severity as i32,
-        peer_address: String::new(),
-        previous_peer_address: String::new(),
-        prefix: String::new(),
-        prefix_length: 0,
-        afi_safi: proto::AddressFamily::Unspecified as i32,
-        summary: format!(
-            "dataplane {} status changed: installed={} rejected={} failed={}",
-            summary.source, summary.installed, summary.rejected, summary.failed
-        ),
-        payload: Some(proto::bgp_event::Payload::Dataplane(payload)),
-    }
-}
-
-fn spawn_dataplane_poller(
-    tx: broadcast::Sender<proto::BgpEvent>,
-    shared_tx: DataplaneEventBroadcaster,
-    blackhole_snapshot: BlackholeDiscardSnapshotFn,
-    fib_snapshot: FibRouteSnapshotFn,
-) {
-    tokio::spawn(async move {
-        let mut previous = dataplane_summaries(&blackhole_snapshot(), &fib_snapshot());
-        let mut interval = tokio::time::interval(DATAPLANE_EVENT_POLL_INTERVAL);
-        interval.set_missed_tick_behavior(tokio::time::MissedTickBehavior::Skip);
-        loop {
-            interval.tick().await;
-            if tx.receiver_count() == 0 {
-                let mut guard = shared_tx
-                    .lock()
-                    .expect("dataplane event broadcaster mutex poisoned");
-                if tx.receiver_count() == 0 {
-                    *guard = None;
-                    return;
-                }
-            }
-
-            let current = dataplane_summaries(&blackhole_snapshot(), &fib_snapshot());
-            for summary in changed_dataplane_summaries(&previous, &current) {
-                let _ = tx.send(dataplane_summary_to_bgp_event(summary));
-            }
-            previous = current;
-        }
-    });
-}
-
-fn parse_watch_events_filter(req: &proto::WatchEventsRequest) -> Result<WatchEventsFilter, Status> {
-    let categories = parse_category_filter(&req.categories)?;
-    let event_types = parse_event_type_filter(&req.event_types)?;
-    let afi = route_event_afi_filter(req.afi_safi)?;
-    let prefix = parse_route_event_prefix_filter(&req.prefix, req.prefix_length, afi)?;
-    let peer = if req.neighbor_address.is_empty() {
-        None
-    } else {
-        Some(
-            req.neighbor_address
-                .parse::<IpAddr>()
-                .map_err(|e| Status::invalid_argument(format!("invalid address: {e}")))?,
-        )
-    };
-
-    Ok(WatchEventsFilter {
-        categories,
-        event_types,
-        peer,
-        afi,
-        prefix,
-    })
-}
-
-fn parse_list_session_events_filter(
-    req: &proto::ListSessionEventsRequest,
-) -> Result<SessionEventHistoryFilter, Status> {
-    let event_types = parse_session_event_type_filter(&req.event_types)?;
-    let peer = if req.neighbor_address.is_empty() {
-        None
-    } else {
-        Some(
-            req.neighbor_address
-                .parse::<IpAddr>()
-                .map_err(|e| Status::invalid_argument(format!("invalid address: {e}")))?,
-        )
-    };
-
-    Ok(SessionEventHistoryFilter {
-        peer,
-        event_types,
-        limit: req.limit as usize,
-    })
-}
-
-fn parse_list_policy_events_filter(
-    req: &proto::ListPolicyEventsRequest,
-) -> Result<PolicyEventHistoryFilter, Status> {
-    parse_policy_event_type_filter(&req.event_types)?;
-    let peer = if req.neighbor_address.is_empty() {
-        None
-    } else {
-        Some(
-            req.neighbor_address
-                .parse::<IpAddr>()
-                .map_err(|e| Status::invalid_argument(format!("invalid address: {e}")))?,
-        )
-    };
-
-    Ok(PolicyEventHistoryFilter {
-        peer,
-        limit: req.limit as usize,
-    })
-}
-
-fn route_event_to_bgp_event(event: rustbgpd_rib::RouteEvent) -> proto::BgpEvent {
-    let event_type = route_event_type_to_bgp_event_type(event.event_type);
-    let route = route_event_to_proto(event);
-    let summary = format!(
-        "route {} {}/{}",
-        match event_type {
-            proto::BgpEventType::RouteAdded => "added",
-            proto::BgpEventType::RouteWithdrawn => "withdrawn",
-            proto::BgpEventType::RouteBestChanged => "best changed",
-            proto::BgpEventType::Unspecified
-            | proto::BgpEventType::SessionStateChanged
-            | proto::BgpEventType::SessionEstablished
-            | proto::BgpEventType::SessionLost
-            | proto::BgpEventType::PeerEnabled
-            | proto::BgpEventType::PeerDisabled
-            | proto::BgpEventType::NotificationSent
-            | proto::BgpEventType::NotificationReceived
-            | proto::BgpEventType::PolicyChanged
-            | proto::BgpEventType::DataplaneStatusChanged
-            | proto::BgpEventType::StreamLagged => "changed",
-        },
-        route.prefix,
-        route.prefix_length
-    );
-
-    proto::BgpEvent {
-        timestamp: route.timestamp.clone(),
-        category: proto::EventCategory::Route as i32,
-        event_type: event_type as i32,
-        severity: proto::EventSeverity::Info as i32,
-        peer_address: route.peer_address.clone(),
-        previous_peer_address: route.previous_peer_address.clone(),
-        prefix: route.prefix.clone(),
-        prefix_length: route.prefix_length,
-        afi_safi: route.afi_safi,
-        summary,
-        payload: Some(proto::bgp_event::Payload::Route(route)),
-    }
-}
-
-fn session_event_to_bgp_event(event: SessionEvent) -> proto::BgpEvent {
-    match event {
-        SessionEvent::Lifecycle(event) => session_lifecycle_event_to_bgp_event(event),
-        SessionEvent::Notification(event) => session_notification_event_to_bgp_event(event),
-    }
-}
-
-fn session_lifecycle_event_to_bgp_event(event: SessionLifecycleEvent) -> proto::BgpEvent {
-    let event_type = session_lifecycle_event_type_to_bgp_event_type(event.event_type);
-    let peer_address = event.peer.to_string();
-    let old_state = event
-        .old_state
-        .map_or_else(String::new, |state| state.as_str().to_string());
-    let new_state = event
-        .new_state
-        .map_or_else(String::new, |state| state.as_str().to_string());
-    let session_role = event.session_role.unwrap_or_default();
-    let session = proto::SessionEvent {
-        event_type: event_type as i32,
-        peer_address: peer_address.clone(),
-        timestamp: event.timestamp.clone(),
-        old_state,
-        new_state,
-        session_role,
-        reason: event.reason.clone(),
-    };
-
-    let severity = match event.event_type {
-        SessionLifecycleEventType::Lost => proto::EventSeverity::Warning,
-        SessionLifecycleEventType::StateChanged
-        | SessionLifecycleEventType::Established
-        | SessionLifecycleEventType::PeerEnabled
-        | SessionLifecycleEventType::PeerDisabled => proto::EventSeverity::Info,
-    };
-
-    proto::BgpEvent {
-        timestamp: event.timestamp,
-        category: proto::EventCategory::Session as i32,
-        event_type: event_type as i32,
-        severity: severity as i32,
-        peer_address,
-        previous_peer_address: String::new(),
-        prefix: String::new(),
-        prefix_length: 0,
-        afi_safi: proto::AddressFamily::Unspecified as i32,
-        summary: event.reason,
-        payload: Some(proto::bgp_event::Payload::Session(session)),
-    }
-}
-
-fn session_notification_event_to_bgp_event(event: SessionNotificationEvent) -> proto::BgpEvent {
-    let event_type = session_notification_event_type_to_bgp_event_type(event.event_type);
-    let peer_address = event.peer.to_string();
-    let direction = match event.event_type {
-        SessionNotificationEventType::Sent => "sent",
-        SessionNotificationEventType::Received => "received",
-    };
-    let notification = proto::NotificationEvent {
-        event_type: event_type as i32,
-        peer_address: peer_address.clone(),
-        timestamp: event.timestamp.clone(),
-        direction: direction.to_string(),
-        code: u32::from(event.code),
-        subcode: u32::from(event.subcode),
-        description: event.description.clone(),
-        session_role: event.session_role.unwrap_or_default(),
-        shutdown_reason: event.shutdown_reason.unwrap_or_default(),
-        reason: event.reason.clone(),
-    };
-
-    proto::BgpEvent {
-        timestamp: event.timestamp,
-        category: proto::EventCategory::Session as i32,
-        event_type: event_type as i32,
-        severity: proto::EventSeverity::Warning as i32,
-        peer_address,
-        previous_peer_address: String::new(),
-        prefix: String::new(),
-        prefix_length: 0,
-        afi_safi: proto::AddressFamily::Unspecified as i32,
-        summary: event.reason,
-        payload: Some(proto::bgp_event::Payload::Notification(notification)),
-    }
-}
-
-fn policy_event_to_bgp_event(event: PolicyEvent) -> proto::BgpEvent {
-    let PolicyEvent {
-        operation,
-        target_type,
-        target,
-        peer,
-        affected_peer_count,
-        timestamp,
-        reason,
-    } = event;
-    let peer_address = peer.map_or_else(String::new, |peer| peer.to_string());
-    let policy = proto::PolicyEvent {
-        event_type: proto::BgpEventType::PolicyChanged as i32,
-        operation: operation.to_string(),
-        target_type: target_type.to_string(),
-        target,
-        peer_address: peer_address.clone(),
-        affected_peer_count: u32::try_from(affected_peer_count).unwrap_or(u32::MAX),
-        timestamp: timestamp.clone(),
-        reason: reason.clone(),
-    };
-
-    proto::BgpEvent {
-        timestamp,
-        category: proto::EventCategory::Policy as i32,
-        event_type: proto::BgpEventType::PolicyChanged as i32,
-        severity: proto::EventSeverity::Info as i32,
-        peer_address,
-        previous_peer_address: String::new(),
-        prefix: String::new(),
-        prefix_length: 0,
-        afi_safi: proto::AddressFamily::Unspecified as i32,
-        summary: reason,
-        payload: Some(proto::bgp_event::Payload::Policy(policy)),
-    }
-}
-
-fn stream_lag_bgp_event(
-    source_category: proto::EventCategory,
-    missed_count: u64,
-) -> proto::BgpEvent {
-    let source = match source_category {
-        proto::EventCategory::Route => "route",
-        proto::EventCategory::Session => "session",
-        proto::EventCategory::Policy
-        | proto::EventCategory::Dataplane
-        | proto::EventCategory::Unspecified => "unknown",
-    };
-    let summary = format!("{source} event stream lagged; missed {missed_count} event(s)");
-    proto::BgpEvent {
-        timestamp: rustbgpd_rib::event::unix_timestamp_now(),
-        category: source_category as i32,
-        event_type: proto::BgpEventType::StreamLagged as i32,
-        severity: proto::EventSeverity::Warning as i32,
-        peer_address: String::new(),
-        previous_peer_address: String::new(),
-        prefix: String::new(),
-        prefix_length: 0,
-        afi_safi: proto::AddressFamily::Unspecified as i32,
-        summary: summary.clone(),
-        payload: Some(proto::bgp_event::Payload::StreamLag(
-            proto::StreamLagEvent {
-                source_category: source_category as i32,
-                missed_count,
-                reason: summary,
-            },
-        )),
     }
 }
 
@@ -1142,13 +391,17 @@ impl proto::event_service_server::EventService for EventService {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::peer_types::{POLICY_EVENT_HISTORY_CAPACITY, SESSION_EVENT_HISTORY_CAPACITY};
+    use crate::peer_types::{
+        POLICY_EVENT_HISTORY_CAPACITY, PolicyEvent, SESSION_EVENT_HISTORY_CAPACITY, SessionEvent,
+        SessionLifecycleEvent, SessionLifecycleEventType, SessionNotificationEvent,
+        SessionNotificationEventType,
+    };
     use crate::proto::event_service_server::EventService as EventServiceTrait;
     use prometheus::Encoder;
     use rustbgpd_fsm::SessionState;
     use rustbgpd_rib::RouteEvent;
     use rustbgpd_wire::{Ipv4Prefix, Ipv6Prefix};
-    use std::net::{Ipv4Addr, Ipv6Addr};
+    use std::net::{IpAddr, Ipv4Addr, Ipv6Addr};
     use std::sync::{Arc, Mutex};
     use tokio::sync::{broadcast, mpsc};
     use tokio_stream::StreamExt;
